refactor: move `ShareSheet` to its own package
