+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/profile/ProfileScreen.kt
@@ -43,7 +43,7 @@ import net.thechance.mena.identity.presentation.screen.profile.components.Profil
 import net.thechance.mena.identity.presentation.screen.profile.components.ShareIcon
 import net.thechance.mena.identity.presentation.screen.notImplemented.NotImplementedScreen
 import net.thechance.mena.identity.presentation.screen.profile.components.share.ShareQrCode
-import net.thechance.mena.identity.presentation.screen.profile.components.bottomSheet.ShareSheet
+import net.thechance.mena.identity.presentation.screen.profile.components.share.ShareSheet
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
